requirements.txt
# LightRAG and RAG-Anything
lightrag-hku[api]>=0.1.0
raganything[all]>=0.1.0

# AI/ML dependencies
openai>=1.0.0,<3.0.0

# Database drivers
neo4j>=5.15.0
psycopg2-binary>=2.9.0
pgvector>=0.2.0

# Web framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0

# Document processing
mineru>=0.1.0
Pillow>=10.0.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
